final cleanup things
