+++ src/pepr/operator/crd/index.ts
@@ -1,6 +1,3 @@
-import { V1OwnerReference } from "@kubernetes/client-node";
-import { GenericKind } from "kubernetes-fluent-client";
-
 export {
   Allow,
   Direction,
@@ -24,16 +21,3 @@ export {
 } from "./generated/exemption-v1alpha1";
 
 export * as Istio from "./generated/istio/virtualservice-v1beta1";
-
-export function getOwnerRef(cr: GenericKind): V1OwnerReference[] {
-  const { name, uid } = cr.metadata!;
-
-  return [
-    {
-      apiVersion: cr.apiVersion!,
-      kind: cr.kind!,
-      uid: uid!,
-      name: name!,
-    },
-  ];
-}
